requirements.txt
klein==15.0.0
twisted==15.0.0
service_identity==14.0.0
jsonschema==2.4.0
iso8601==0.1.8
lxml==3.4.1
treq==0.2.1
silverberg==0.1.12
pyOpenSSL==0.14
jsonfig==0.1.1
testtools==0.9.32
croniter==0.3.5
txkazoo==0.0.6b2
effect==0.10
txeffect==0.9
characteristic==14.3.0
attrs==15.0.0
toolz==0.7.4
pyrsistent==0.10.3
singledispatch==3.4.0.3
six==1.9.0
sumtypes==0.1a4
